dont pull .git files

+++ setup.sh
@@ -11,10 +11,14 @@ fi
 BACKUP_DIR=~/.dotbackup/$(date +%y%m%d%H%M%S)
 mkdir -p $BACKUP_DIR
 for file in ~/.dotconf/.[^.]*; do
-    #Use greedy glob removal to grab the basename of each dotfile.
     BASENAME=${file##*/}
-    mv ~/$BASENAME $BACKUP_DIR/$BASENAME
-    cp -r ~/.dotconf/$BASENAME ~/$BASENAME
+    if [[ $BASENAME != ".git" ]] && [[ $BASENAME != ".gitmodules" ]]; then
+        #Use greedy glob removal to grab the basename of each dotfile.
+        if [ -f ~/.$BASENAME ]; then
+            mv ~/$BASENAME $BACKUP_DIR/$BASENAME
+        fi
+        cp -r ~/.dotconf/$BASENAME ~/$BASENAME
+    fi
 done
 
 echo "Don't forget to source the shell profile!"
